refactor: 유효성 검사를 객체에서 진행하도록 수정
- Validator가 분리되어 있을 때 발생하는 혼동을 막기 위해 도메인 객체에서 유효성 검사를 진행하도록 수정
