+++ src/main/java/domain/Line.java
@@ -1,19 +1,24 @@
 package domain;
 
-import domain.validator.LineValidator;
-
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import java.util.Random;
 
 public class Line {
+
+    public static final int POINT_MIN_SIZE = 0;
+    public static final int POINT_MAX_SIZE = 19;
 
     private final List<Point> points;
 
     public Line(final List<Point> points) {
-        LineValidator.validate(points.size());
+        validate(points.size());
         this.points = points;
+    }
+
+    public void validate(final int pointSize) {
+        if (pointSize < POINT_MIN_SIZE || pointSize > POINT_MAX_SIZE) {
+            throw new IllegalArgumentException("포인트 범위는 0부터 19까지입니다.");
+        }
     }
 
     public Point getPointAt(final int index) {
