--- src/main/java/domain/validator/LineValidator.java
@@ -1,14 +1,0 @@
-package domain.validator;
-
-public class LineValidator {
-
-    public static final int POINT_MIN_SIZE = 0;
-    public static final int POINT_MAX_SIZE = 19;
-
-    public static void validate(final int pointSize) {
-        if (pointSize < POINT_MIN_SIZE || pointSize > POINT_MAX_SIZE) {
-            throw new IllegalArgumentException("포인트 범위는 0부터 19까지입니다.");
-        }
-    }
-
-}
